# src/documentation/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CookLink API Documentation",
    "version": "1.0.0",
    "description": "API documentation for CookLink Capstone Project, by Fatma Zehra Aydın, Tareq Harh, and Omar Qaqish"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "GoogleOAuth": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "http://localhost:3000/api/auth/google",
            "tokenUrl": "https://accounts.google.com/o/oauth2/token",
            "scopes": {
              "https://www.googleapis.com/auth/userinfo.profile": "All user operations requiring authentication."
            }
          }
        }
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 30
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 6
          },
          "passwordChangedAt": {
            "type": "string",
            "format": "date-time"
          },
          "provider": {
            "type": "string"
          },
          "providerId": {
            "type": "string"
          },
          "address": {
            "type": "string",
            "format": "ObjectId"
          },
          "profilePicture": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["admin", "cook", "user"],
            "default": "user"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "location": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "dishes": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "ObjectId"
            }
          }
        },
        "required": ["username", "email"]
      },
...
    }
  },
  "paths": {
...
    "/api/user": {
      "get": {
        "summary": "Get authenticated user's profile",
        "operationId": "getMyInfo",
        "tags": ["User"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                },
                "example": {
                  "username": "string",
                  "firstName": "string",
                  "lastName": "string",
                  "email": "user@example.com",
                  "address": {
                    "street": "string",
                    "buildingNumber": 0,
                    "apartmentNumber": 0,
                    "neighborhood": "string",
                    "district": "string",
                    "city": "string",
                    "country": "string",
                    "postalCode": 0
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Update authenticated user's profile",
        "operationId": "updateUserProfile",
        "tags": ["User"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete authenticated user's account",
        "operationId": "deleteMyAccount",
        "tags": ["User"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Account deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}